Admin rights in card

+++ src/main/java/project/trello/service/CardService.java
@@ -1,6 +1,5 @@
 package project.trello.service;
 
-import org.apache.catalina.User;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import project.trello.ProjectApplication;
@@ -19,6 +18,7 @@ public class CardService {
     private final ArchiveRepository archiveRepository;
     private final UsersRepository usersRepository;
     private final BoardRepository boardRepository;
+    private final UsersService usersService;
 
     @Autowired
     public CardService(CardRepository cardRepository,
@@ -27,7 +27,7 @@ public class CardService {
                        ListRepository listRepository,
                        ArchiveRepository archiveRepository,
                        UsersRepository usersRepository,
-                       BoardRepository boardRepository) {
+                       BoardRepository boardRepository, UsersService usersService) {
         this.cardRepository = cardRepository;
         this.labelRepository = labelRepository;
         this.commentRepository = commentRepository;
@@ -35,22 +35,30 @@ public class CardService {
         this.archiveRepository = archiveRepository;
         this.usersRepository = usersRepository;
         this.boardRepository = boardRepository;
+        this.usersService = usersService;
     }
 
     public project.trello.model.List createCard(Long list_id, Card card) {
-        card.setList_id(list_id);
-        cardRepository.save(card);
-        project.trello.model.List list = listRepository.findById(list_id).get();
-        Card card1 = cardRepository.findById(card.getId()).get();
-        list.getCards().add(card1);
-        Users thisUser = usersRepository.findById(ProjectApplication.user_id).get();
-        String msg = thisUser.getFirstName() + " " + thisUser.getLastName() +
-                " Create card by " + card.getTitle() + " title in " +
-                listRepository.findById(list_id).get().getTitle() + " list.";
+        Long maybeAdmin_id = ProjectApplication.user_id;
         project.trello.model.List thisList = listRepository.findById(list_id).get();
         Long boardId = thisList.getBoard_id();
-        boardRepository.findById(boardId).get().getActivityList().add(msg);
-        return listRepository.save(list);
+        Board thisBoard = boardRepository.findById(boardId).get();
+        Long workspace_id = thisBoard.getWorkspace_id();
+        if (usersService.isAdmin(maybeAdmin_id, workspace_id)) {
+            card.setList_id(list_id);
+            cardRepository.save(card);
+            project.trello.model.List list = listRepository.findById(list_id).get();
+            Card card1 = cardRepository.findById(card.getId()).get();
+            list.getCards().add(card1);
+            Users thisUser = usersRepository.findById(ProjectApplication.user_id).get();
+            String msg = thisUser.getFirstName() + " " + thisUser.getLastName() +
+                    " Create card by " + card.getTitle() + " title in " +
+                    listRepository.findById(list_id).get().getTitle() + " list.";
+
+            boardRepository.findById(boardId).get().getActivityList().add(msg);
+            return listRepository.save(list);
+        }
+        throw new IllegalStateException("You are not an admin :/");
     }
 
     public List<Card> getCards(){
@@ -58,31 +66,38 @@ public class CardService {
     }
 
     public void deleteCard(Long card_id) {
-        boolean exists = cardRepository.existsById(card_id);
-        if (!exists) {
-            throw new IllegalStateException("card with id " + card_id +
-                    " does not exist!");
-        }
-        if(archiveRepository.findAll().isEmpty()){
-            throw new IllegalStateException("archive is empty !");
-        }
-        Archive archive = archiveRepository.findById(1L).get();
-        List<Card> archiveCards = archive.getCards();
-        for(Card card : archiveCards){
-            if(card.getId().equals(card_id)){
-                Users thisUser = usersRepository.findById(ProjectApplication.user_id).get();
-                Long listId = cardRepository.findById(card_id).get().getList_id();
-                project.trello.model.List thisList = listRepository.findById(listId).get();
-                String msg = thisUser.getFirstName() + " " + thisUser.getLastName() +
-                        " delete card by " + card.getTitle() + " title " +
-                        "from " + thisList.getTitle() + " list.";
-                Long boardId = thisList.getBoard_id();
-                boardRepository.findById(boardId).get().getActivityList().add(msg);
-                cardRepository.deleteById(card_id);
-                return;
+        Long maybeAdmin_id = ProjectApplication.user_id;
+        Long list_id = cardRepository.findById(card_id).get().getList_id();
+        project.trello.model.List thisList = listRepository.findById(list_id).get();
+        Long boardId = thisList.getBoard_id();
+        Board thisBoard = boardRepository.findById(boardId).get();
+        Long workspace_id = thisBoard.getWorkspace_id();
+        if (usersService.isAdmin(maybeAdmin_id, workspace_id)) {
+            boolean exists = cardRepository.existsById(card_id);
+            if (!exists) {
+                throw new IllegalStateException("card with id " + card_id +
+                        " does not exist!");
             }
+            if (archiveRepository.findAll().isEmpty()) {
+                throw new IllegalStateException("archive is empty !");
+            }
+            Archive archive = archiveRepository.findById(1L).get();
+            List<Card> archiveCards = archive.getCards();
+            for (Card card : archiveCards) {
+                if (card.getId().equals(card_id)) {
+                    Users thisUser = usersRepository.findById(ProjectApplication.user_id).get();
+                    Long listId = cardRepository.findById(card_id).get().getList_id();
+                    String msg = thisUser.getFirstName() + " " + thisUser.getLastName() +
+                            " delete card by " + card.getTitle() + " title " +
+                            "from " + thisList.getTitle() + " list.";
+                    boardRepository.findById(boardId).get().getActivityList().add(msg);
+                    cardRepository.deleteById(card_id);
+                    return;
+                }
+            }
+            throw new IllegalStateException("you should archive this card first !");
         }
-        throw new IllegalStateException("you should archive this card first !");
+        throw new IllegalStateException("You are not an admin :/");
     }
 
     public Card editCard(Long card_id,Card card){
